Zones controller using versioncake for versioning
