+++ api/app/views/spree/api/zones/index.v1.rabl
@@ -1,0 +1,4 @@
+object false
+child(@zones => :zones) do
+  extends 'spree/api/zones/show'
+end
