+++ api/app/views/spree/api/zones/show.v1.rabl
