+++ api/lib/spree_api.rb
@@ -1,1 +1,2 @@
 require 'spree/api'
+require 'versioncake'
